refactor(components): refactor timeline
